requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-braces==1.14.0
django-chartjs==2.1.0
django-highcharts==0.1.7
idna==2.9
pyaml==20.4.0
pyswagger==0.8.39
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
validate-email==1.3
